code beautify
ProcessMemoryUtil関係を別ファイルへ移した
